requirements/common.txt
Django~=4.1.3
pillow
djangorestframework
django-cors-headers